chore: fix all clippy warnings (workspace)
Brings the workspace to clippy-clean on stable 1.87.0 so the CI lint
job can be lifted from advisory to required in the next commit.

Changes by class:
- map_or simplifications: mechanical rewrite via clippy --fix
- complex-type aliases: SegmentPartitionCache, SegmentPartitionSlice,
  DeconvResult, and RankKeptCtx struct in
  crates/scoring/src/scoring/scored_spectrum.rs
- too_many_arguments: RankKeptCtx context struct in scored_spectrum.rs;
  #[allow] with reason for directional_node_score_inner, write_tsv,
  write_tsv_to, write_spectrum_rows, and compute_cleavage_credit
- doc-list indentation: add blank line after list / fix continuation
  indent at 15 sites in msgf-rust.rs and scored_spectrum.rs
- unused_mut, ? rewrite, manual split_once, loop-counter: via clippy --fix
- needless_range_loop: suppressed with reason (seg indexes cache AND
  serves as fallback partition_for arg)

No functional behavior change; PIN/TSV bit-identical regression gate
in tree (precursor_cal_bit_identical) is the verification.

+++ crates/scoring/tests/add_prob_dist_chunked_parity.rs
@@ -31,8 +31,8 @@ fn add_prob_dist_scalar(
     for t in t_start..t_end {
         let src_idx = (t - other_min) as usize;
         let dst_idx = (t + score_diff - self_min) as usize;
-        let cur = dst.get_probability((t + score_diff) as i32);
-        dst.set_prob((t + score_diff) as i32, cur + src_p(src, src_idx) * aa_prob);
+        let cur = dst.get_probability((t + score_diff));
+        dst.set_prob((t + score_diff), cur + src_p(src, src_idx) * aa_prob);
         let _ = dst_idx; // silence
     }
 }
